lambda/requirements.txt
numpy==1.19.2
pillow==7.2.0
